Fixed dropudp and findstring. Deleted string as it seemed to be a incomplete duplicate of findstring.
