--- benchmarks/programs/dropudp/compress/size.txt
@@ -1,1 +1,0 @@
-original,754
