+++ benchmarks/programs/dropudp/prog.c
@@ -14,7 +14,7 @@ static int IFINDEX;
  * @brief Unload the eBPF program from the XDP and
  */
 void _unloadProg() {
-    bpf_set_link_xdp_fd(IFINDEX, -1, XDPFLAGS);
+    bpf_xdp_attach(IFINDEX, -1, XDPFLAGS, NULL);
     printf("Unloading the eBPF program...");
     exit(0);
 }
@@ -23,26 +23,30 @@ void _unloadProg() {
  * @brief Load the eBPF program    
  */
 int main(int argc, char **argv) {
-    int success, prog_fd;
+    int prog_fd, success;
     struct bpf_object *obj;
+    struct bpf_program *prog;    
+
     obj = bpf_object__open("prog.bpf.o");
+
+    prog = bpf_object__next_program(obj, NULL);
+    bpf_program__set_type(prog, BPF_PROG_TYPE_XDP);
+
     success = bpf_object__load(obj);
+
+    // Load the program
+    if (success != 0) {
+        printf("The kernel didn't load the BPF program\n");
+        return -1;
+    }
 
     signal(SIGINT, _unloadProg);
     signal(SIGTERM, _unloadProg);
-
-    if (success != 0) {
-        printf("Error loading the eBPF program\n");
-        return -1;
-    }
     
-    struct bpf_program *prog;
 
-    prog = bpf_object__find_program_by_name(obj, PROGNAME);
     prog_fd = bpf_program__fd(prog);
-
     IFINDEX = if_nametoindex("lo");
-    if (bpf_set_link_xdp_fd(IFINDEX, prog_fd, XDPFLAGS) < 0) {
+    if (bpf_xdp_attach(IFINDEX, prog_fd, XDPFLAGS, NULL) < 0) {
         printf("link set xdp fd failed\n");
         return -1;
     }
