--- benchmarks/programs/string/Makefile
@@ -1,1 +1,0 @@
--include ../Makefile.config
